# Fast lexer