--- a/compiler/lib/src/main/resources/META-INF/native-image/reflect-config.json
+++ b/compiler/lib/src/main/resources/META-INF/native-image/reflect-config.json
@@ -3762,14 +3762,6 @@
   "fields":[{"name":"0bitmap$1"}]
 },
 {
-  "name":"fpp.compiler.syntax.Lexer$",
-  "fields":[{"name":"0bitmap$2"}]
-},
-{
-  "name":"fpp.compiler.syntax.Lexer$$anon$1",
-  "fields":[{"name":"0bitmap$1"}]
-},
-{
   "name":"fpp.compiler.syntax.Parser$",
   "fields":[{"name":"0bitmap$1"}]
 },