--- a/compiler/lib/src/main/scala/analysis/Analysis.scala
+++ b/compiler/lib/src/main/scala/analysis/Analysis.scala
@@ -261,7 +261,7 @@ case class Analysis(
     val Value.Integer(v) = Analysis.convertValueToType(
       valueMap(id),
       Type.Integer
-    )
+    ): @unchecked
     v
   }