--- a/compiler/lib/src/main/scala/codegen/CppWriter/TlmPacketSetCppWriter.scala
+++ b/compiler/lib/src/main/scala/codegen/CppWriter/TlmPacketSetCppWriter.scala
@@ -225,8 +225,8 @@ case class TlmPacketSetCppWriter(
     lines(
       s"""|
           |// The size of the array of omitted channels
-          |constexpr FwIndexType omittedArraySize = $size;""",
-    ),
+          |constexpr FwIndexType omittedArraySize = $size;"""
+    )
   }
 
   private def writeOmittedChannels: List[Line] = {