--- a/compiler/lib/src/main/scala/util/File.scala
+++ b/compiler/lib/src/main/scala/util/File.scala
@@ -1,5 +1,8 @@
 package fpp.compiler.util
 
+import java.io.{ByteArrayOutputStream, FileReader}
+import java.nio.file.Files
+
 /** A file used in compilation */
 sealed trait File {
 
@@ -9,22 +12,22 @@ sealed trait File {
   }
 
   /** Open the file for reading */
-  def openRead(locOpt: Option[Location] = None): Result.Result[java.io.Reader] = {
+  def openRead(locOpt: Option[Location] = None): Result.Result[Array[Char]] = {
     try {
       val r = this match {
-        case File.Path(p) => new java.io.FileReader(p.toFile)
-        case File.StdIn => new java.io.BufferedReader(new java.io.InputStreamReader(System.in))
+        case File.Path(p) => scala.io.Source.fromFile(p.toFile)
+        case File.StdIn => scala.io.Source.fromInputStream(System.in)
       }
-      Right(r)
+
+      Right(r.toArray)
     }
     catch {
-      case _: Exception => Left(FileError.CannotOpen(locOpt, this.toString))
+      case e: Exception => Left(FileError.CannotOpen(locOpt, this.toString))
     }
   }
 
   /** Open the file for writing */
   def openWrite(locOpt: Option[Location] = None): Result.Result[java.io.PrintWriter] = {
-    val error = FileError.CannotOpen(locOpt, this.toString)
     try {
       this match {
         case File.Path(p) => {
@@ -33,11 +36,12 @@ sealed trait File {
           val printWriter = new java.io.PrintWriter(fileWriter)
           Right(printWriter)
         }
-        case _ => Left(error)
+        case _ => Left(FileError.CannotOpen(locOpt, this.toString)
+        )
       }
     }
     catch {
-      case _: Exception => Left(error)
+      case e: Exception => Left(FileError.CannotOpen(locOpt, this.toString))
     }
   }
 
@@ -66,11 +70,13 @@ object File {
       val prefix = java.nio.file.Paths.get(s)
       if (path.startsWith(prefix)) prefix.relativize(path) else path
     }
+
     prefixes.map(removePrefix(_)) match {
       case Nil => path
       case head :: tail => {
-        def min(p1: JavaPath, p2: JavaPath) = 
+        def min(p1: JavaPath, p2: JavaPath) =
           if (p1.getNameCount < p2.getNameCount) p1 else p2
+
         tail.fold(head)(min)
       }
     }