--- a/compiler/project/build.properties
+++ b/compiler/project/build.properties
@@ -1,1 +1,1 @@
-sbt.version=1.6.2
+sbt.version=1.10.11