--- a/compiler/project/scala3-migration.sbt
+++ b/compiler/project/scala3-migration.sbt
@@ -1,2 +1,0 @@
-// https://docs.scala-lang.org/scala3/guides/migration/scala3-migrate.html
-addSbtPlugin("ch.epfl.scala" % "sbt-scala3-migrate" % "0.5.0")