This PR reimplements the lexing strategy that produces a hugely faster lexer (8-10x speedup). It uses the same design pattern as can be found in the Scala compiler using a manually written recursive descent parser that moves through the file character-by-character. I also included some fixes in the parser that speed up the parsing by ~30% using IntelliJ's static analysis and auto-fixes. Mainly the fixes removed some redunant syntax and added `private` to functions that could be internalized (allows inlining and other optimizations by Scala).

## Notable changes

There are some very minor discrepancies between the old and new lexer, @bocchino I didn't get exact feature parity because some behavior seems a little under defined so I'll wait for your input:

1. Trailing spaces after a `\` line continuation are disallowed. There is a single file in the FPrime repo that has some trailing spaces which throws an error in the lexer. I'm leaning toward keeping the functionality as is since it enforces cleaner syntax.
2. The lexer uses an error `Context` to track invalid tokens or tokenization errors.
3.  The string trimming of multi-line strings (3-strings) is pretty confusing (or bugged). The old behavior had odd behavior:

```
constant F: string = """
    indented
    keeps
    trailing newline
    """
```

Generates a string with value: `indented\nkeeps\ntrailing newline\n`

```
constant F: string = """
non-indented
does not keep
trailing newline
```

Generates string with value: `non-indented\ndoes not keep\ntrailing newline`

For now the new lexer will not string trailing newlines ever but I can easily add that in.

## Benchmarks

To benchmark I ran a JVM-based profiler to compare lexing times on the locs generation step of the FppTest build (this is a parsing intensive build stage). The images screenshots of the runtime of the lexer before and after this PR. I also `time`ed the runtime of the native GraalVM build of `fpp-locate-defs` before and after this PR. The stats are included below.

Before:
![image](https://github.com/user-attachments/assets/61bc0f98-fbfe-46ff-8581-2e9e2573119d)

```
real	0m3.460s
user	0m3.389s
sys	0m0.071s
```

After:
![image](https://github.com/user-attachments/assets/341b690e-35b5-47b0-ac1b-87c5cab094d4)
```
real	0m0.335s
user	0m0.284s
sys	0m0.052s
```