@Kronos3 2025-04-29
I have pushed the following changes
- Clarified whitespace handling of multi-line string literals in the spec. The old behavior did not match the spec. We decided to update the spec to match the behavior.
- Updated lexer to allow trailing spaces after a line continuation, any other character will trigger an error
- Added a testcase for line continuation and multi-error reporting (updated `illegal-character` test in `fpp-syntax`)