@Kronos3 2025-04-29
More interesting benchmarks from our CI jobs:

Before
<img width="1364" alt="Screenshot 2025-04-29 at 11 34 00 AM" src="https://github.com/user-attachments/assets/73267a1d-4119-4452-93cf-4f4d1b949ab4" />

After
<img width="1364" alt="Screenshot 2025-04-29 at 11 33 52 AM" src="https://github.com/user-attachments/assets/39a1f004-b253-4805-b61d-77e6dac9cd81" />

Before
<img width="1346" alt="Screenshot 2025-04-29 at 12 47 04 PM" src="https://github.com/user-attachments/assets/d20c1e0f-4cc1-4a24-bb73-fb753e96ab0d" />

After
<img width="1346" alt="image" src="https://github.com/user-attachments/assets/1b4d1637-c301-4914-95fd-6f53b9dc4959" />