@Kronos3 2025-05-07
I neglected to put the most important benchmark here which is the cmake configuration time. I purged FppTest before each run and ran a native GraalVM build before/after:

Before:
```
-- Configuring done (23.9s)
-- Generating done (1.7s)
```

After:
```
-- Configuring done (10.4s)
-- Generating done (1.7s)
```